requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
